refactor(cache): split AST cache responsibilities
